feat: add scatter operator, distinguish tensor and communication APIs via namespaces, and reorganize functions in misc files
